preliminary compatibility with bsd/darwin

+++ zsh/zshrc.d/01_alias.zsh
@@ -1,14 +1,22 @@
 # Aliases
 alias cd=' cd'
-alias ls=' ls -F --group-directories-first --color=auto'
+if  [[ $OSTYPE =~ gnu ]] 
+then
+    alias ls=' ls -FH --group-directories-first --color=auto'
+    alias mv='nocorrect noglob timeout 5 mv -iv'
+    alias rm='nocorrect noglob timeout 5 rm -Iv --one-file-system'
+elif [[ $OSTYPE =~ (darwin|bsd) ]]
+then
+    alias ls=' ls -FGH'
+    alias mv='nocorrect noglob mv -iv'
+    alias rm='nocorrect noglob rm -iv'
+fi
 alias ll='ls -l'
 alias la='ls -A'
 alias lla='ll -A'
 alias lh='ll -h'
 alias lah='lla -h'
 alias :q='exit'
-alias mv='nocorrect noglob timeout 5 mv -iv'
-alias rm='nocorrect noglob timeout 5 rm -Iv --one-file-system'
 alias cp="cp -i"                          # confirm before overwriting something
 alias grep='grep -i --color=tty -d skip'
 alias df='df -h'                          # human-readable sizes
